+++ src/osu/mod.rs
@@ -1,1 +1,2 @@
+pub mod osu2016;
 pub mod osu2018;